+++ test/specs/shallow-mount.spec.js
@@ -1,6 +1,6 @@
 import { compileToFunctions } from 'vue-template-compiler'
 import Vue from 'vue'
-import { mount, shallowMount } from '~vue/test-utils'
+import { mount, shallowMount, createLocalVue } from '~vue/test-utils'
 import Component from '~resources/components/component.vue'
 import ComponentWithChild from '~resources/components/component-with-child.vue'
 import ComponentWithNestedChildren from '~resources/components/component-with-nested-children.vue'
@@ -254,6 +254,23 @@ describeRunIf(process.env.TEST_ENV !== 'node', 'shallowMount', () => {
 
     expect(wrapper.find(ChildComponent).vm.propA)
       .to.equal('hey')
+  })
+
+  it('stubs components registered on localVue after multiple installs', () => {
+    const myPlugin = function (_Vue, opts) {
+      _Vue.mixin({ })
+    }
+    const localVue = createLocalVue()
+    localVue.component('registered-component', {
+      render: h => h('time')
+    })
+    const TestComponent = {
+      render: h => h('registered-component')
+    }
+
+    localVue.use(myPlugin)
+    const wrapper = shallowMount(TestComponent, { localVue })
+    expect(wrapper.html()).to.contain('registered-component-stub')
   })
 
   it('throws an error when the component fails to mount', () => {